requirements.txt
faker
dbt-bigquery
google-cloud-bigquery
numpy
pandas
pyarrow
